+++ gke/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2018-09-17
+
+- Disable istio egress gateway by default
+- Able is enable/disable istio control pane
+- Able is enable/disable istio egress gateway
+- Able is enable/disable kubernetes-dashboard
+
 ## 2018-09-16
 
 - Install addon `kubernetes-dashboard` in namespace `kubernetes-dashboard`